requirements.txt
arrow==1.1.1
astroid==2.6.5
feedparser==6.0.8
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.9.6
python-dateutil==2.8.2
python-dotenv==0.19.0
rfeed==1.1.1
sgmllib3k==1.0.0
six==1.16.0
toml==0.10.2
wrapt==1.12.1
